refactor: Product - MenuProduct 간 객체참조 제거
