+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -1,12 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Entity
@@ -16,32 +14,25 @@ public class MenuProduct {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "seq")
     private Long id;
-    @ManyToOne
-    @JoinColumn(name = "product_id")
-    private Product product;
+    @Column
+    private Long productId;
     @Column(name = "quantity")
     private long quantity;
-    @ManyToOne
-    private Menu menu;
 
     public MenuProduct() {
     }
 
-    public MenuProduct(Product product, long quantity) {
-        this.product = product;
+    public MenuProduct(Long productId, long quantity) {
+        this.productId = productId;
         this.quantity = quantity;
-    }
-
-    public void changeMenu(Menu menu) {
-        this.menu = menu;
     }
 
     public Long getId() {
         return id;
     }
 
-    public Product getProduct() {
-        return product;
+    public Long getProductId() {
+        return productId;
     }
 
     public long getQuantity() {
