+++ src/main/java/kitchenpos/domain/menu/MenuProducts.java
@@ -1,0 +1,27 @@
+package kitchenpos.domain.menu;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embeddable;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import java.util.List;
+
+@Embeddable
+public class MenuProducts {
+
+    @OneToMany(cascade = CascadeType.ALL)
+    @JoinColumn(name = "menu_id", nullable = false)
+    private List<MenuProduct> menuProducts;
+
+    public MenuProducts() {
+    }
+
+    public MenuProducts(List<MenuProduct> menuProducts) {
+        this.menuProducts = menuProducts;
+    }
+
+    public List<MenuProduct> getMenuProducts() {
+        return menuProducts;
+    }
+
+}
